requirements.txt
os
logging
python-dotenv
python-telegram-bot
google-generative-ai
functools32  
nltk
geopy
